Rename all snake_cased folders within src/components/elements/reactions
