+++ app/packs/src/components/elements/reactions/ReactionDetails.js
@@ -12,11 +12,11 @@ import ElementActions from 'src/stores/alt/actions/ElementActions';
 import DetailActions from 'src/stores/alt/actions/DetailActions';
 import LoadingActions from 'src/stores/alt/actions/LoadingActions';
 import ReactionDetailsLiteratures from 'src/components/elements/DetailsTabLiteratures';
-import ReactionDetailsContainers from 'src/components/elements/reactions/analyses_tab/ReactionDetailsContainers';
+import ReactionDetailsContainers from 'src/components/elements/reactions/analysesTab/ReactionDetailsContainers';
 import SampleDetailsContainers from 'src/components/elements/samples/analyses_tab/SampleDetailsContainers';
-import ReactionDetailsScheme from 'src/components/elements/reactions/scheme_tab/ReactionDetailsScheme';
-import ReactionDetailsProperties from 'src/components/elements/reactions/properties_tab/ReactionDetailsProperties';
-import GreenChemistry from 'src/components/elements/reactions/green_chemistry_tab/GreenChemistry';
+import ReactionDetailsScheme from 'src/components/elements/reactions/schemeTab/ReactionDetailsScheme';
+import ReactionDetailsProperties from 'src/components/elements/reactions/propertiesTab/ReactionDetailsProperties';
+import GreenChemistry from 'src/components/elements/reactions/greenChemistryTab/GreenChemistry';
 import Utils from 'src/utilities/Functions';
 import PrintCodeButton from 'src/components/common/PrintCodeButton';
 import XTabs from 'src/components/extra/ReactionDetailsXTabs';
