+++ app/packs/src/components/elements/reactions/analysesTab/ReactionDetailsContainers.js
