+++ app/packs/src/components/elements/reactions/propertiesTab/ReactionDetailsProperties.js
