+++ app/packs/src/components/elements/reactions/schemeTab/ReactionDetailsDuration.js
